fix(visual-editing): use VisualEditing React component directly (#183)
* fix(deps): update @sanity/visual-editing

* chore: .gitignore *.tgz

* fix(deps): update @sanity/visual-editing

* fix(deps): remove sanity deps from dependencies

Astro seems to only work properly when these are treated as true
`peerDependenceis`. Which means that users of this library need to make sure
they have them installed in their own app.

* fix(visual-editing): use VisualEditing React component directly

This change removes the need for `enableVisualEditing()` and bypasses the
lazy-loading. The lazy-loading runs React rendering internally which does not
seem to play well with Astro.